cleanup

+++ product-service/src/main/java/com/eyamattoussi/productservice/dto/ProductRequestDto.java
@@ -4,9 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.Id;
-
-import java.math.BigDecimal;
 @Data
 @Builder
 @AllArgsConstructor
@@ -14,5 +11,5 @@ import java.math.BigDecimal;
 public class ProductRequestDto {
     private String name;
     private String description;
-    private BigDecimal price;
+    private Double price;
 }
