+++ product-service/src/main/java/com/eyamattoussi/productservice/dto/ProductResponseDto.java
@@ -4,9 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.Id;
-
-import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -16,5 +13,5 @@ public class ProductResponseDto {
     private String id;
     private String name;
     private String description;
-    private BigDecimal price;
+    private Double price;
 }
