+++ product-service/src/main/java/com/eyamattoussi/productservice/model/Product.java
@@ -4,8 +4,6 @@ import lombok.*;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-import java.math.BigDecimal;
-
 @Document(value = "product")
 @Builder
 @AllArgsConstructor
@@ -18,5 +16,6 @@ public class Product {
     private String id;
     private String name;
     private String description;
-    private BigDecimal price;
+    private Double price;
+
 }
